+++ src/view/ProfileItem.tsx
@@ -3,15 +3,15 @@ import { profileContext } from '../reader/profile'
 import PlayerReader from '../reader/player'
 import { useContext } from 'react'
 import { gameContext } from '../reader/game'
+import Curtain from './Curtain'
 
 export default function ProfileItemView (): JSX.Element {
   const profileState = useContext(profileContext)
   const gameState = useContext(gameContext)
-  const player = gameState.phase !== 'join' && <PlayerReader />
+  const playing = gameState.phase !== 'join'
   return (
-    <>
+    <Curtain open={playing} openElement={<PlayerReader />}>
       <Text>Profile: {profileState.userId}</Text>
-      {player}
-    </>
+    </Curtain>
   )
 }
